requirements.txt
openai
transformers
langchain
pydantic>=2.0.0
pydantic-settings>=2.0.0
gymnasium
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
requests>=2.31.0
slowapi>=0.1.9
aiohttp>=3.8.0
pyyaml>=6.0
python-multipart>=0.0.6
psutil>=5.9.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
python-dotenv>=1.0.0